+++ cms/service/EvaluationService.py
@@ -41,11 +41,10 @@ from sqlalchemy.exc import IntegrityError
 from cms import ServiceCoord
 from cmscommon.datetime import make_timestamp
 from cms.db import SessionGen, Digest, Dataset, Evaluation, Submission, \
-    SubmissionResult, Testcase, UserTest, UserTestResult, get_submissions, \
-    get_submission_results, get_datasets_to_judge
+    SubmissionResult, Testcase, UserTest, UserTestResult, get_datasets_to_judge
 from cms.grading.Job import Job
 from cms.io import Service, rpc_method
-from .esoperations import ESOperation, get_relevant_operations, \
+from .esoperations import ESOperation, \
     submission_get_operations, submission_to_evaluate, \
     user_test_get_operations
 
@@ -684,118 +683,3 @@ class EvaluationService(Service):
             self.enqueue_all(self.get_user_test_operations(user_test))
 
             session.commit()
-
-    @rpc_method
-    @with_post_finish_lock
-    def invalidate_submission(self,
-                              contest_id=None,
-                              submission_id=None,
-                              dataset_id=None,
-                              participation_id=None,
-                              task_id=None,
-                              level="compilation"):
-        """Request to invalidate some computed data.
-
-        Invalidate the compilation and/or evaluation data of the
-        SubmissionResults that:
-        - belong to submission_id or, if None, to any submission of
-          participation_id and/or task_id or, if both None, to any
-          submission of the contest asked for, or, if all three are
-          None, the contest this service is running for (or all contests).
-        - belong to dataset_id or, if None, to any dataset of task_id
-          or, if None, to any dataset of any task of the contest this
-          service is running for.
-
-        The data is cleared, the operations involving the submissions
-        currently enqueued are deleted, and the ones already assigned to
-        the workers are ignored. New appropriate operations are
-        enqueued.
-
-        submission_id (int|None): id of the submission to invalidate,
-            or None.
-        dataset_id (int|None): id of the dataset to invalidate, or
-            None.
-        participation_id (int|None): id of the participation to
-            invalidate, or None.
-        task_id (int|None): id of the task to invalidate, or None.
-        level (string): 'compilation' or 'evaluation'
-
-        """
-        logger.info("Invalidation request received.")
-
-        # Validate arguments
-        # TODO Check that all these objects belong to this contest.
-        if level not in ("compilation", "evaluation"):
-            raise ValueError(
-                "Unexpected invalidation level `%s'." % level)
-
-        if contest_id is None:
-            contest_id = self.contest_id
-
-        with SessionGen() as session:
-            # When invalidating a dataset we need to know the task_id,
-            # otherwise get_submissions will return all the submissions of
-            # the contest.
-            if dataset_id is not None and task_id is None \
-                    and submission_id is None:
-                task_id = Dataset.get_from_id(dataset_id, session).task_id
-            # First we load all involved submissions.
-            submissions = get_submissions(
-                session,
-                # Give contest_id only if all others are None.
-                contest_id
-                if {participation_id, task_id, submission_id} == {None}
-                else None,
-                participation_id, task_id, submission_id).all()
-
-            # Then we get all relevant operations, and we remove them
-            # both from the queue and from the pool (i.e., we ignore
-            # the workers involved in those operations).
-            operations = get_relevant_operations(
-                level, submissions, dataset_id)
-            for operation in operations:
-                try:
-                    self.queue_service.dequeue(operation=operation.to_dict())
-                except KeyError:
-                    pass  # Ok, the operation wasn't in the queue.
-                try:
-                    self.queue_service.ignore_operation(
-                        operation=operation.to_dict()
-                    )
-                except LookupError:
-                    pass  # Ok, the operation wasn't in the pool.
-
-            # Then we find all existing results in the database, and
-            # we remove them.
-            submission_results = get_submission_results(
-                session,
-                # Give contest_id only if all others are None.
-                contest_id
-                if {participation_id,
-                    task_id,
-                    submission_id,
-                    dataset_id} == {None}
-                else None,
-                participation_id,
-                # Provide the task_id only if the entire task has to be
-                # reevaluated and not only a specific dataset.
-                task_id if dataset_id is None else None,
-                submission_id, dataset_id).all()
-            logger.info("Submission results to invalidate %s for: %d.",
-                        level, len(submission_results))
-            for submission_result in submission_results:
-                # We invalidate the appropriate data and queue the
-                # operations to recompute those data.
-                if level == "compilation":
-                    submission_result.invalidate_compilation()
-                elif level == "evaluation":
-                    submission_result.invalidate_evaluation()
-
-            # Finally, we re-enqueue the operations for the
-            # submissions.
-            for submission in submissions:
-                self.enqueue_all(self.get_submission_operations(submission))
-
-            session.commit()
-        logger.info("Invalidate successfully completed.")
-
